Split interop into interop-cats,interop-future and interop-scalaz7x.
