+++ build.sbt
@@ -30,7 +30,21 @@ lazy val root = project
     skip in publish := true,
     console := (console in Compile in coreJVM).value
   )
-  .aggregate(coreJVM, coreJS, interopJVM, interopJS, interopCatsLaws, benchmarks, microsite)
+  .aggregate(
+    coreJVM,
+    coreJS,
+    interopJVM,
+    interopJS,
+    interopCatsJVM,
+    interopCatsJS,
+    interopFutureJVM,
+    interopFutureJS,
+    interopScalaz7xJVM,
+    interopScalaz7xJS,
+    interopCatsLaws,
+    benchmarks,
+    microsite
+  )
   .enablePlugins(ScalaJSPlugin)
 
 lazy val core = crossProject(JSPlatform, JVMPlatform)
@@ -85,12 +99,6 @@ lazy val interop = crossProject(JSPlatform, JVMPlatform)
   .settings(stdSettings("zio-interop"))
   .dependsOn(core % "test->test;compile->compile")
   .settings(
-    libraryDependencies ++= Seq(
-      "org.scalaz"    %%% "scalaz-core"               % "7.2.+" % Optional,
-      "org.typelevel" %%% "cats-effect"               % "1.0.0" % Optional,
-      "org.scalaz"    %%% "scalaz-scalacheck-binding" % "7.2.+" % Test,
-      "co.fs2"        %%% "fs2-core"                  % "1.0.0" % Test
-    ),
     scalacOptions in Test ++= Seq("-Yrangepos")
   )
 
@@ -98,12 +106,56 @@ lazy val interopJVM = interop.jvm
 
 lazy val interopJS = interop.js
 
+lazy val interopCats = crossProject(JSPlatform, JVMPlatform)
+  .in(file("interop-cats"))
+  .settings(stdSettings("zio-interop-cats"))
+  .dependsOn(core % "test->test;compile->compile")
+  .settings(
+    libraryDependencies ++= Seq(
+      "org.typelevel" %%% "cats-effect" % "1.0.0" % Optional,
+      "co.fs2"        %%% "fs2-core"    % "1.0.0" % Test
+    ),
+    scalacOptions in Test ++= Seq("-Yrangepos")
+  )
+
+lazy val interopCatsJVM = interopCats.jvm.dependsOn(interopJVM)
+
+lazy val interopCatsJS = interopCats.js.dependsOn(interopJS)
+
+lazy val interopFuture = crossProject(JSPlatform, JVMPlatform)
+  .in(file("interop-future"))
+  .settings(stdSettings("zio-interop-future"))
+  .dependsOn(core % "test->test;compile->compile")
+  .settings(
+    scalacOptions in Test ++= Seq("-Yrangepos")
+  )
+
+lazy val interopFutureJVM = interopFuture.jvm.dependsOn(interopJVM)
+
+lazy val interopFutureJS = interopFuture.js.dependsOn(interopJS)
+
+lazy val interopScalaz7x = crossProject(JSPlatform, JVMPlatform)
+  .in(file("interop-scala7x"))
+  .settings(stdSettings("zio-interop-scala7x"))
+  .dependsOn(core % "test->test;compile->compile")
+  .settings(
+    libraryDependencies ++= Seq(
+      "org.scalaz" %%% "scalaz-core"               % "7.2.+" % Optional,
+      "org.scalaz" %%% "scalaz-scalacheck-binding" % "7.2.+" % Test
+    ),
+    scalacOptions in Test ++= Seq("-Yrangepos")
+  )
+
+lazy val interopScalaz7xJVM = interopScalaz7x.jvm.dependsOn(interopJVM)
+
+lazy val interopScalaz7xJS = interopScalaz7x.js.dependsOn(interopJS)
+
 // Separated due to binary incompatibility in scalacheck 1.13 vs 1.14
 // TODO remove it when https://github.com/typelevel/discipline/issues/52 is closed
 lazy val interopCatsLaws = project.module
   .in(file("interop-cats-laws"))
   .settings(stdSettings("zio-interop-cats-laws"))
-  .dependsOn(coreJVM % "test->test", interopJVM % "test->compile")
+  .dependsOn(coreJVM % "test->test", interopJVM % "test->compile", interopCatsJVM % "test->compile")
   .settings(
     skip in publish := true,
     libraryDependencies ++= Seq(
