+++ interop-scala7x/jvm/src/main/scala/scalaz/zio/interop/console/scalaz.scala
@@ -1,0 +1,21 @@
+package scalaz
+package zio
+package interop
+package console
+
+import zio.{ console => c }
+
+object scalaz {
+
+  /**
+   * Prints the string representation of an object to the console.
+   */
+  def putStr[A](a: A)(implicit ev: Show[A]): IO[Nothing, Unit] =
+    c.putStr(ev.shows(a))
+
+  /**
+   * Prints the string representation of an object to the console, including a newline character.
+   */
+  def putStrLn[A](a: A)(implicit ev: Show[A]): IO[Nothing, Unit] =
+    c.putStrLn(ev.shows(a))
+}
